Move all configs to invoke_training/config.
