+++ src/invoke_training/config/shared/training_output_config.py
