+++ src/invoke_training/scripts/invoke_textual_inversion_sd.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import yaml
 
-from invoke_training.training.config.textual_inversion_config import (
+from invoke_training.config.pipelines.textual_inversion_config import (
     TextualInversionConfig,
 )
 from invoke_training.training.textual_inversion.textual_inversion_sd import run_training
