+++ src/invoke_training/training/finetune_lora/finetune_lora_sdxl.py
@@ -16,11 +16,11 @@ from diffusers.optimization import get_scheduler
 from tqdm.auto import tqdm
 from transformers import CLIPPreTrainedModel, PretrainedConfig, PreTrainedTokenizer
 
+from invoke_training.config.pipelines.finetune_lora_config import FinetuneLoRASDXLConfig
 from invoke_training.core.lora.injection.stable_diffusion import (
     inject_lora_into_clip_text_encoder,
     inject_lora_into_unet,
 )
-from invoke_training.training.config.finetune_lora_config import FinetuneLoRASDXLConfig
 from invoke_training.training.finetune_lora.tokenize_captions import tokenize_captions
 from invoke_training.training2.shared.accelerator.accelerator_utils import (
     get_mixed_precision_dtype,
